netdb: dns: no need to proceed if not initialized
In dns_find_answer(), if it is invoked before dns_bind(), it blocks
in sem_take because dns not initialized hence sem_init hasn't been
called.

And it seems reasonable to skip in dns_find_answer() if dns not
initialized, since this means we might bring up network for the
first time, nothing has been recorded ever.